+++ src/main/java/com/darksci/pardot/api/response/emailclick/EmailClickQueryResponse.java
@@ -1,0 +1,66 @@
+/**
+ * Copyright 2017 Stephen Powis https://github.com/Crim/pardot-java-client
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated
+ * documentation files (the "Software"), to deal in the Software without restriction, including without limitation the
+ * rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and to permit
+ * persons to whom the Software is furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all copies or substantial portions of the
+ * Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE
+ * WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
+ * COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR
+ * OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ */
+
+package com.darksci.pardot.api.response.emailclick;
+
+import com.fasterxml.jackson.dataformat.xml.annotation.JacksonXmlProperty;
+
+import java.util.List;
+
+/**
+ * Represents the results of querying Email Clicks in Pardot API.
+ */
+public class EmailClickQueryResponse {
+    private EmailClickQueryResponse.Result result;
+
+    public Result getResult() {
+        return result;
+    }
+
+    /**
+     * Represents the results of querying Campaigns in Pardot API.
+     */
+    public static class Result {
+        private Integer totalResults = 0;
+
+        @JacksonXmlProperty(localName = "emailClick")
+        private List<EmailClick> emailClicks;
+
+        public Integer getTotalResults() {
+            return totalResults;
+        }
+
+        public List<EmailClick> getEmailClicks() {
+            return emailClicks;
+        }
+
+        @Override
+        public String toString() {
+            return "Result{"
+                + "totalResults=" + totalResults
+                + ", emailClicks=" + emailClicks
+                + '}';
+        }
+    }
+
+    @Override
+    public String toString() {
+        return "EmailClickQueryResponse{"
+            + "result=" + result
+            + '}';
+    }
+}